version 0.1/backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Werkzeug
PyJWT
Flask-Cors
Flask-Mail